+++ src/interfaces/index.ts
@@ -1,6 +1,7 @@
 export { MiddlewareFn, NextFn, MiddlewareInterface } from "./Middleware";
 export * from "./AuthChecker";
 export * from "./ClassType";
+export * from "./Complexity";
 export * from "./Publisher";
 export * from "./ResolverData";
 export * from "./ResolverFilterData";