+++ src/metadata/definitions/field-metadata.ts
@@ -1,6 +1,7 @@
 import { ParamMetadata } from "./param-metadata";
 import { TypeValueThunk, TypeOptions } from "../../decorators/types";
 import { Middleware } from "../../interfaces/Middleware";
+import { Complexity } from "../../interfaces";
 
 export interface FieldMetadata {
   target: Function;
@@ -10,6 +11,7 @@ export interface FieldMetadata {
   typeOptions: TypeOptions;
   description: string | undefined;
   deprecationReason: string | undefined;
+  complexity: Complexity | undefined;
   params?: ParamMetadata[];
   roles?: any[];
   middlewares?: Array<Middleware<any>>;