+++ src/metadata/definitions/resolver-metadata.ts
@@ -7,12 +7,14 @@ import {
 } from "../../decorators/types";
 import { ParamMetadata } from "./param-metadata";
 import { Middleware } from "../../interfaces/Middleware";
+import { Complexity } from "../../interfaces";
 
 export interface BaseResolverMetadata {
   methodName: string;
   schemaName: string;
   target: Function;
   handler: Function | undefined;
+  complexity: Complexity | undefined;
   resolverClassMetadata?: ResolverClassMetadata;
   params?: ParamMetadata[];
   roles?: any[];