+++ src/metadata/metadata-storage.ts
@@ -197,6 +197,7 @@ export class MetadataStorage {
             typeOptions: def.typeOptions!,
             deprecationReason: def.deprecationReason,
             description: def.description,
+            complexity: def.complexity,
             roles: def.roles!,
             middlewares: def.middlewares!,
             params: def.params!,
@@ -204,6 +205,7 @@ export class MetadataStorage {
           this.collectClassFieldMetadata(fieldMetadata);
           objectType.fields!.push(fieldMetadata);
         } else {
+          objectTypeField.complexity = def.complexity;
           if (objectTypeField.params!.length === 0) {
             objectTypeField.params = def.params!;
           }